refactor(types): shared common types
