+++ plugins/qeta-backend/src/database/QetaStore.ts
@@ -2,60 +2,11 @@ import {
   Statistic,
   StatisticsRequestParameters,
 } from '@drodil/backstage-plugin-qeta-common';
-
-export interface Question {
-  id: number;
-  author: string;
-  title: string;
-  content: string;
-  created: Date;
-  updated?: Date;
-  updatedBy?: string;
-  score: number;
-  views: number;
-  answersCount: number;
-  correctAnswer: boolean;
-  favorite: boolean;
-  ownVote?: number;
-  tags?: string[];
-  entities?: string[];
-  answers?: Answer[];
-  own?: boolean;
-  votes?: Vote[];
-  trend?: number;
-  comments?: Comment[];
-}
-
-export interface Answer {
-  id: number;
-  questionId: number;
-  author: string;
-  content: string;
-  correct: boolean;
-  created: Date;
-  updated?: Date;
-  updatedBy?: string;
-  score: number;
-  ownVote?: number;
-  own?: boolean;
-  votes?: Vote[];
-  comments?: Comment[];
-}
-
-export interface Vote {
-  author: string;
-  score: number;
-  timestamp: Date;
-}
-
-export interface Comment {
-  author: string;
-  content: string;
-  created: Date;
-  own?: boolean;
-  updated?: Date;
-  updatedBy?: string;
-}
+import type {
+  Question,
+  Answer,
+  Attachment,
+} from '@drodil/backstage-plugin-qeta-common';
 
 export type MaybeAnswer = Answer | null;
 export type MaybeQuestion = Question | null;
@@ -96,18 +47,7 @@ export interface TagResponse {
   tag: string;
   questionsCount: number;
 }
-export interface Attachment {
-  id: number;
-  uuid: string;
-  locationType: string;
-  locationUri: string;
-  path: string;
-  binaryImage: Buffer;
-  mimeType: string;
-  extension: string;
-  creator: string;
-  created: Date;
-}
+
 export interface AttachmentParameters {
   uuid: string;
   locationType: string;
