+++ plugins/qeta/src/api/types.ts
@@ -1,4 +1,9 @@
 import { ErrorObject } from 'ajv';
+import {
+  Question,
+  Answer,
+  Attachment,
+} from '@drodil/backstage-plugin-qeta-common';
 
 interface CustomError {
   message: string;
@@ -10,45 +15,13 @@ interface ErrorResponse {
 }
 
 export interface QuestionsResponse {
-  questions: QuestionResponse[];
+  questions: Question[];
   total: number;
-}
-
-export interface CommentResponse {
-  id: number;
-  content: string;
-  author: string;
-  created: Date;
-  own: boolean;
-  updated?: Date;
-  updatedBy?: string;
 }
 
 export type QuestionsResponseBody = QuestionsResponse | ErrorResponse;
 
-export interface QuestionResponse {
-  id: number;
-  author: string;
-  title: string;
-  content: string;
-  tags?: string[];
-  entities?: string[];
-  created: Date;
-  views: number;
-  score: number;
-  answersCount: number;
-  correctAnswer: boolean;
-  favorite: boolean;
-  ownVote?: number;
-  updated?: string;
-  updatedBy?: string;
-  own?: boolean;
-  answers?: AnswerResponse[];
-  votes?: VoteResponse[];
-  comments: CommentResponse[];
-}
-
-export type QuestionResponseBody = QuestionResponse | ErrorResponse;
+export type QuestionResponseBody = Question | ErrorResponse;
 
 export interface QuestionRequest {
   title: string;
@@ -64,46 +37,14 @@ export interface AnswerRequest {
   images?: number[];
 }
 
-export interface AnswerResponse {
-  id: number;
-  author: string;
-  content: string;
-  questionId: number;
-  created: Date;
-  score: number;
-  updated?: Date;
-  correct: boolean;
-  updatedBy: string;
-  ownVote?: number;
-  own?: boolean;
-  votes?: VoteResponse[];
-  comments?: CommentResponse[];
-}
-
-export type AnswerResponseBody = AnswerResponse | ErrorResponse;
-
-export interface VoteResponse {
-  author: string;
-  score: number;
-  timestamp: Date;
-}
+export type AnswerResponseBody = Answer | ErrorResponse;
 
 export interface TagResponse {
   tag: string;
   questionsCount: number;
 }
 
-export interface AttachmentResponse {
-  id: number;
-  uuid: string;
-  locationType: string;
-  locationUri: string;
-  path: string;
-  binaryImage: Buffer;
-  mimeType: string;
-  extension: string;
-  creator: string;
-  created: Date;
-}
+export type AttachmentResponseBody = Attachment | ErrorResponse;
 
-export type AttachmentResponseBody = AttachmentResponse | ErrorResponse;
+export type QuestionResponse = Question;
+export type AnswerResponse = Answer;
